xous: definitions: dispose of Vec on hosted
When running on Hosted mode, reconstitute the Vec and dispose of it
rather than freeing the memory by passing it back to the kernel.

This is because the data was previously allocated by way of a vec, and
only a reconstituted vec knows how to properly dispose of the data.

Signed-off-by: Sean Cross <sean@xobs.io>