improve resilience to ping flood attack
much better, but still eventually the stack can fall over eventually.
also, the stack doesn't behave well when you try to initiate new
commands in the middle of a flood, because the back-pressure mechanisms
are primarily in-stack at the moment.

I think i have an idea on how to fix that but this will have to be later.