detect when the WF200 goes catatonic and reset it
This is not the fix I was hoping for for connection stability,
but unfortunately this is where we are at.

The problem is that at some point the WF200 throws a "host interface
error"; but the error codes show no specific error. After throwing
this error, it can respond to some basic commands (such as the one
asking what the current status and error codes are), but higher level
APIs seem to time out. There's no command to bring the WF200 stack
to a sane state (like a warm reset or buffer flush) as far as I
can see, so, the solution is to detect the issue and send a reset
to the WF200. This keeps the link up in the long term, but
you will experience drop-outs about once every 2-3 hours for
about 20 seconds while the system detects and fixes the problem.

I think for this one, more data will be needed from customer units
to see if this could be just a local network issue with my config
or a more pervasive problem.