Sync the EC to the SOC on boot
Always reset the EC on boot. If we have a cold-boot situation,
it's hard to know what state the EC is in versus our stack;
a hard reset on the EC guarantees we at least always start from
the same point.

Note that on suspend/resume, there is no reset - the EC continues
trundling onwards, but the system was designed for this.