take advantage of signed length in the kernel
this prevents us from having to use the entire reserved region
as a worst-case and improves screen-up time on resume.

previously, we didn't encode a signed length in the kernel so we couldn't
trust the length that was presented to us. This was fixed a while ago
but i just connected the dots today that this needed to be updated as well.