seems like the DH key exchange is broken somehow? needs more digging.
i've extracted the key parameters from the test suite and confirmed that
that the hash value is computed correctly on their side.

i've also confirmed that the hash value is computed correctly from
the locally computed key.

so what's gone wrong is the devired secret key doesn't match what's
on the other side.

possible issues could be either a PK serialization issue, or something
wrong in the DH exchange itself.