fix forward-version autoupdate flow and change boot ordering
in order to trigger an update in early boot, all the security
sensitive servers have to have their connection tables occupied
(I guess the fancy term is the "capabilities" need to be allocated.
whatevs. #notanacademic).

This fixes the flow so those capabilities are occupied before
we get to the update check. Also fixes options to ignore updates.