prep a fork of riscv package
for now, publish a xous-riscv package, until we can absorb
this into the kernel directly.

Want to recode all the assembly files into inline-assembly now
that it is supported in Rust, but, let's not put that yak on
the list of yaks to shave before we have a separable kernel.