fix visibility issues in API
API now has to be shared with the crate that implements the
API, so they are now all public structures.

Not exactly great, but, probably also not exactly harmful because
all the source is open and anyone looking to inject messages
can just find the source code and construct them manually.

Mostly we just have to be careful to not abuse the API by reaching
around the lib calls and passing raw xous messages based on
what could be changing data structures.