renode: remove custom LiteX_Timer_32
When we started this project, all Renode Litex modules were 8-bits, and
we had to custom-make a 32-bit timer module. However, in the interim,
upstream has grown a 32-bit variant of this timer. As of this writing,
this change was made two years ago.

Remove our custom variant and use the upstream version instead.

Signed-off-by: Sean Cross <sean@xobs.io>