add automounter to shellchat, and initial wait message
The initial wait message is drawn directly under where
the modals would appear, so the message "goes away"
when modals pop up.

There's still a bit of trickiness around the timing of
some of the calls. Attempting to mount the PDDB too early
before all the contexts have been registered leads to
a lot of needless defacing.

Also, the final redraw needs a short delay to allow
the system to foreground shellchat again after closing
the modals.