add a table-based implementation of cosine
this is a space-optimized implementation of cosine. it's about
12k smaller than core lib's "proper" cosine.

Gives about 9 bits amplitude resolution, and divides the waveform
into 1024 time positions. Good enough for generating tones on
a crappy speaker.

Later on if anyone needs to "really" do cosine or sine, we'll have
a footgun for them, but...let's hope Rust fixes issue
https://github.com/rust-lang/rust/issues/105734 before then, and
we can deprecate this library in favor of the mathematically
accurate and canonical cosine implementation.