work around some more RTC derpiness
This isn't in the spec sheet, and I can't figure out why it
is the case, but emperically it is the case.

The RTC is much more reliable at absorbing values writen
to it if they are immediately read back.

I suspect it may have something to do with their internal
synchronizer that locks out updates to registers while
any of the time registers are being updated. It doesn't say
that a read is required to clear that lock, but empricially
sometimes I find the RTC gets "hung" returning 0xFF all
the time, unless some dummy reads follow up writes.