fix regression due to #366
Boxing the `Codec` object and sticking it on the heap causes
audio playback to hard-lock (no kp, no other explanation given).

Not sure what the issue is, but a cleaner fix is arguably to simply
increase the stack size so the objects can be allocated there.

I suspect the problem may have to do with how the FIFO structures
are memory-mapped and somehow this does not play nice with
Box (there are `unsafe` routines within it).