fix trust level of framebuffer canvas
this is some bit-rot that happened to the `ball` demo because
it's currently the only application that uses the framebuffer
canvas.

the trust level of the framebuffer canvas has to be lower than
that of the modals, otherwise a modals box is not allowed to
draw on top of the framebuffer.

in this case, a patch that was applied when trust levels were
refactored was missed on the framebuffer canvas, which caused
the trust level of the canvas to be higher than that of a modals.

Thus, when a modal attempts to pop up over the ball demo,
one just sees the defacement lines but no actal content.