add routines to set/get EC readiness
The nominal purpose of this is to allow other services to delay
interactions that could interfere with an EC update.

Note that anyone could call this routine and mark the EC as not
ready, since the llio block is not secured and the API is public.

I don't think it's a security problem, but is a DoS vector.