add `std` as a target for engine-sha512
we actually do have a `std` now in Xous, it wasn't there before
because we made this crate before we had `std`.

However, I'm not quite sure what apis we should be adding if we
have `std`...probably something else should be added, we'll probably
trip over it when someone tries to compile against a crate that
actually uses the `std`-specific APIs.