format services/root-keys
formatted some pretty big blocks of either autogenerated
or copy/pasted blobs used for emulation.

I think I'll let the formatting stand because the autogen
code is "never" going to change (i.e. they are physical constants
tied to the properties of the hardware) or the pasted blobs
have no real meaning beyond the blob itself so if they were
regenerated it'd be another copypasta with format.