change order of imports (work around rust-analyzer bug?)
I can't figure out why rust-analyzer on my system isn't work
correctly but if I change the order of imports things work better.

If this breaks things for Foundation Devices (@eupn) let me know,
I will poke at this some more and try to figure it out. In theory
it should have no impact but somehow I'm getting the atsamd target
*always* live in my vscode environment, even when it is switched
off when it isn't the last target in the list (and this overrides
all of the rust-analyzer settings for other targets).