add RawTrng SysCall to Xous API
This SysCall is introduced because not all hardware targets can
afford the luxury of a private TRNG port for the kernel.

In this case, an API is needed to exfiltrate random numbers from
the kernel so we can inspect them and check them for health.

If a TRNG is operating correctly, cherry-picking some values
out of it to QC the operation of the TRNG should be harmless.

Of course, nothing is truly stateless, and there is always a chance
that someone could construct a gadget using this to reduce the
potential state-space of the final random number generator. Therefore,
this API is (a) gated behind a feature flag and (b) any implementation
that plans to expose this should be coded with some countermeasures
to reduce data leakage after sampling.