fix broken bio test
the fifo_basic() test case - if not allowed to run for a while -
can lead to some strange lockups in the system fifos. it's
an interesting question as to why the fifo's can't be "cleaned up"
in a sane fashion when it gets scrambled, but for now, the answer
is I think "don't do that".