xtask: exit after printing help
Causes the xtask build to print an error and exit if it doesn't
recognize the task argument. Just ran into this as a result of the
cramsoc -> bao1x rename in 46d828b -- the build wasn't recognizing
`baremetal-cramsoc` anymore but was proceeding with the compile anyway,
without setting the target or loader features.