add 'dabao' series of apps
these are apps that are meant to be targeting "headless"
dabao style configurations. the decision was made to split
'apps' and 'apps-dabao' because for sure people are going
to try to load 'apps' onto a a 'dabao' board and wonder
why they don't work.

And the reason is the dabao just doesn't have the features
of a fully-featured system. so the 'apps-dabao' hopefully gives
people a clue that there's a difference between the two types
of apps.

also a very simple hello world app is added.