Add a script for automating the signing of bao1x images.
Image signing has to happen on the native host computer and not
in any of the VMs so a powershell script was created to allow
automation of the 'beta' signing key. This script fully automates
the signing and also staging of images to the testing environment.

This will need to be revised into a linux script for final
deployment as CI is done in a linux environment, but "day to day"
signature testing is done in the development environment, thus
this version of the script was prioritized.

Shouldn't be too hard to extend this to the linux environment,
I think Windows is the harder one to integrate.