deny boot to any developer-key signed kernel/detached-app/swap images
Secret keys can't be erased after boot1 exits because the protection
bits are hardened from modification after that point.

thus it is not possible for the loader to switch the device
into developer mode.

Thus the only way to enter developer mode is through a dev-mode
signed loader. A system that has never been put into developer
mode will refuse to run any developer-mode xous kernel, app, swap
images.

Once a system has been exposed to a dev-mode signed loader,
the boot1 will flip the developer mode bit and erase the secret
keys.

Later on, if a production loader is replaced, it doesn't matter
that a production loader is loading dev-mode signed successors
because the secret keys are already erased.

Note that in all cases, the key revocations take precedence
so if the developer key is revoked, these cases should never
even be encountered.