harden boot1
boot1 hardening is similar to boot0 hardening - bollard & delay
insertion, checking that we're in PLL mode. Also the first phase
mesh results are checked and second phase is setup for the loader
to finally check.

clocks are still local to the routine - there's a to-do to consolidate
that routine into the API layer later on.