add multi-pass validation in paranoid mode to boot0
this patch uses the previously modified APIs and implements
a multi-pass check on signatures in the case that paranoid
mode is turned on. This may effectively double the boot time but
it also reduces the chance of a successful glitch attack
on signature validation by a multiple of the glitch
success probability.

this also updates other API calls and implements the
XOR-on-target-address code.