remove superfluous prints in boot0
removing prints in boot0 gives less handle for glitchers to time
off of. Also makes debugging chips harder, but I think the boot0
is starting to get stable enough that this is OK?

also removed the frac-PLL setting because it's broken. this causes
the PLL in the boot0 to be subtly different from boot1, and I
think I like it - it makes the timing of glitches between the
two phases different.